+++ src/gr/gr_mod_base.h
@@ -57,7 +57,7 @@ public slots:
     void start(int buffer_size=0);
     void stop();
     int set_data(std::vector<u_int8_t> *data);
-    void tune(long long center_freq);
+    void tune(long center_freq);
     void set_power(float value, std::string gain_stage="");
     void set_filter_width(int filter_width, int mode);
     void set_ctcss(float value);
